project_structure/requirements.txt
# Core Data & ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.0.0
tensorflow==2.14.0
torch==2.0.0
transformers==4.33.0
peft==0.4.0  # For LoRA/QLoRA

# Vector DB & embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
chromadb==0.4.0

# API & Web
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.1.0
sqlalchemy==2.0.21
alembic==1.12.1

# Orchestration & ETL
apache-airflow==2.7.1
apache-airflow-providers-mysql==3.4.1

# Monitoring & Experiment Tracking
mlflow==2.9.0
prometheus-client==0.18.0
wandb==0.16.0

# Local LLM (Ollama integration via API)
requests==2.31.0
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.0

# Testing & Quality
pytest==7.4.2
pytest-cov==4.1.0
black==23.10.0
flake8==6.1.0
isort==5.12.0
mypy==1.6.0

# Drift Detection
evidently==0.4.18

# Development
ipython==8.16.0
jupyter==1.0.0
notebook==7.0.5
